fix: voila-vuetify stalls on "Executing cell x of x"
Changing the requirejs mapping after the nodeps version of
jupyter-vue is loaded, is too late. Some times jupyter-vuetify is
loaded before jupyter-vue. Setting the CDN path directly before
loading solves this at the expense of duplicating the CDN URL
generation from the Voila loader.

Specifying the CDN URL as fallback for [moduleName]/nodeps is to
maintain the same behavior as the Voila loader. This allows to
load local versions of the modules for testing purposes.