fix: scientific notation of Y-axis not shown on first load
Sometimes the scientific notation for the Y-axis is not applied.
When it happens, it's resolved by resizing the browser window,
which causes a re-render of bqplot, so looks to me like bqplot is
sometimes not re-rendering when the tick_format is changed.

To work around this, set the tick_format before the plot is
rendered with the first data.

See: #865