feat: use Solara FileBrowser to import files and directories.
This file browser is more in line with the rest of the look and feel
(not based on ipywidgets-controls, but on ipyvuetify).

Also, looking at the comments at #573 this is implemented using:

   * Single click to select a directory or file
   * Import button for directory or file
   * Double click to quickly import a file.

Tooltips are added to make clear what each button does, or why
it is disabled.

This requires solara >= 0.3