fix: custom styles work in solara server (#2658)
In #2580 we fixed support for custom style in combination with
ipypopout. However, this uses a global widget, which does not work
with solara, since multiple virtual kernels would then share the same
widget.

This is using a singleton decorator that make sure each virtual kernel
will create only 1 StyleRegistry widget.