allow importers to provide list of viewer classes to ignore
when trying to load into viewers.  This effectively allows viewers to _support_ loading data of a certain type, but the importer to still have priority for creating a viewer of a different type.  lcviz will need to use this for TPF cubes which should default to a cube viewer even if a light curve viewer exists (which technically can also display the TPF)