HDUListToSpectrumMixin > spectrum_common.SpectrumInputExtensionsMixin
also renames input_hdulist to input_has_extensions

this is to prepare for further generalization to support 1d spectra as well as extensions from ASDF files.  We may in the future need to have a flag to tell HDUList vs ASDF here, but I think existing traitlets used in the classes really only care if there are extensions or not.